Override Generalizer::generalizeQuantity() to calculate monthly quantitny correctly
Before this change, actions that represent monthly resources, but have come multiplier (e.g. rack_unit)
were transformed to chrges, and charges were  generalized to bill with the same quantity as charge.

It means that if server has 4 units, qty of charge will be exactly 4, so the bill had the same qty. It was
misleading because the bill itself is for 1 month, not for 4.

After this change, monthly charges are ALWAYS generalized to bill with qty equal to period of month,
action belongs to.